safari-600.7-branch/LayoutTests/css3/viewport-percentage-lengths/viewport-percentage-lengths-relative-font-size-expected.txt

Test that relative font sizes work when the parent is in viewport units.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS innerWidth is 800
PASS innerHeight is 600
PASS getComputedStyle(container).fontSize is "40px"
PASS getComputedStyle(ems).fontSize is "160px"
PASS getComputedStyle(percent).fontSize is "20px"
PASS innerWidth is 900
PASS innerHeight is 600
PASS getComputedStyle(container).fontSize is "45px"
PASS getComputedStyle(ems).fontSize is "180px"
PASS getComputedStyle(percent).fontSize is "23px"
PASS innerWidth is 900
PASS innerHeight is 640
PASS getComputedStyle(container).fontSize is "45px"
PASS getComputedStyle(ems).fontSize is "180px"
PASS getComputedStyle(percent).fontSize is "23px"
PASS innerWidth is 500
PASS innerHeight is 640
PASS getComputedStyle(container).fontSize is "25px"
PASS getComputedStyle(ems).fontSize is "100px"
PASS getComputedStyle(percent).fontSize is "13px"
PASS innerWidth is 800
PASS innerHeight is 600
PASS getComputedStyle(container).fontSize is "40px"
PASS getComputedStyle(ems).fontSize is "160px"
PASS getComputedStyle(percent).fontSize is "20px"
